+++ src/main/java/com/nexa/api/catalogmanagement/presentation/rest/response/CatalogMediaResponse.java
@@ -1,0 +1,3 @@
+package com.nexa.api.catalogmanagement.presentation.rest.response;
+
+public record CatalogMediaResponse(String url, String fileName) { }