+++ src/main/java/com/nexa/api/catalogmanagement/presentation/rest/response/MoneyResponse.java
@@ -1,0 +1,3 @@
+package com.nexa.api.catalogmanagement.presentation.rest.response;
+
+public record MoneyResponse(String amount, String currency) { }